Reworking asserts into static assert
Where static assert could not be made (one occasion) I replaced it with
if THROW. Note that THROW does not make sense outside APDU handling.
Thus using it in menu was an error. This is why I decided agains the use
of ASSERT in this case. In future, it may be good to add this, but it
needs a more robust solution as one gound in our cardano app.