Update makefile - speculos tests
The aim is to minimize the risk of build passing despite the failure. In
original Makefile node result was piped and errors were detected by a
script checking the logs. This removes the script altogether. Every
possible error node finds compiling/running the tests will be reported
and will cause make to fail. This is done while keeping most convenient
printouts.